jwt token added for auth

+++ controllers/notes.go
@@ -1,32 +1,27 @@
 package controllers
 
 import (
-	"log"
+	"fmt"
 	"net/http"
 	"notesApp/configs"
 	"notesApp/schemas"
-	"notesApp/validators"
 	"strconv"
 
 	"github.com/gin-gonic/gin"
 )
 
 func FetchNotes(ctx *gin.Context) {
-	// get and validate session id
-	incomingUser := &schemas.User{}
-	// if err := ctx.ShouldBindJSON(incomingUser); err != nil {
-	// 	ctx.JSON(http.StatusBadRequest, gin.H{
-	// 		"error": err.Error(),
-	// 	})
-	// 	return
-	// }
-	// TO DO
-
-	requestUser, _ := ctx.Get("user")
+	// get user email sent by validateJwtToken middleware
+	userEmail, isPresent := ctx.Get("userEmail")
+	if !isPresent {
+		ctx.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
 
 	// fetch all notes
 	getNotes := []schemas.Notes{}
-	result := configs.DefaultDB.Where("email = ?", requestUser.Email).Find(&getNotes)
+	email := fmt.Sprintf("%v", userEmail)
+	result := configs.DefaultDB.Where("email = ?", email).Find(&getNotes)
 	if result.Error != nil {
 		ctx.JSON(http.StatusBadRequest, gin.H{
 			"error": result.Error.Error(),
@@ -39,7 +34,6 @@ func FetchNotes(ctx *gin.Context) {
 }
 
 func CreateNotes(ctx *gin.Context) {
-	notes := &schemas.Notes{}
 	incomingNotes := &schemas.IncomingNotes{}
 	// binding notes json to struct
 	if err := ctx.ShouldBindJSON(incomingNotes); err != nil {
@@ -48,11 +42,16 @@ func CreateNotes(ctx *gin.Context) {
 		})
 		return
 	}
-	// get and validate session id
-	email := validators.EmailExtractor(incomingNotes.Sid)
+	// get user email sent by validateJwtToken middleware
+	userEmail, isPresent := ctx.Get("userEmail")
+	if !isPresent {
+		ctx.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
 
 	// create notes
-	notes.Email = email
+	notes := &schemas.Notes{}
+	notes.Email = fmt.Sprintf("%v", userEmail)
 	notes.Note = incomingNotes.Note
 	result := configs.DefaultDB.Create(&notes)
 	if result.Error != nil {
@@ -67,7 +66,6 @@ func CreateNotes(ctx *gin.Context) {
 }
 
 func DeleteNotes(ctx *gin.Context) {
-	sid := ctx.Param("sid")
 	nid, err := strconv.ParseUint(ctx.Param("nid"), 10, 32)
 	if err != nil {
 		// Handle error if the parameter cannot be converted to a uint32
@@ -77,9 +75,22 @@ func DeleteNotes(ctx *gin.Context) {
 		return
 	}
 
+	// get user email sent by validateJwtToken middleware
+	_, isPresent := ctx.Get("userEmail")
+	if !isPresent {
+		ctx.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+
 	notes := &schemas.Notes{}
-	notes.Email = email
 	notes.NID = uint32(nid)
+	findNid := configs.DefaultDB.First(&notes)
+	if findNid.Error != nil {
+		ctx.JSON(http.StatusBadRequest, gin.H{
+			"error": findNid.Error.Error(),
+		})
+		return
+	}
 	result := configs.DefaultDB.Delete(&notes)
 	if result.Error != nil {
 		ctx.JSON(http.StatusBadRequest, gin.H{
@@ -87,9 +98,8 @@ func DeleteNotes(ctx *gin.Context) {
 		})
 		return
 	}
-	log.Println("deleted rows", result.RowsAffected)
 	ctx.JSON(http.StatusOK, gin.H{
-		"message": notes,
+		"message": "success",
 	})
 
 }
